Refactored CodeAnalysis part to own classlib and introduced Compilation abstraction over the binder
