+++ .editorconfig
@@ -1,0 +1,5 @@
+[*]
+end_of_line = crlf
+
+[*.xml]
+indent_style = space
