+++ Minsk.sln
@@ -3,7 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "mc", "mc\mc.csproj", "{A5E27C6E-623B-46D2-86F5-4127E4A95283}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Minsk.ConsoleApp", "src\app\Minsk.ConsoleApp\Minsk.ConsoleApp.csproj", "{A5E27C6E-623B-46D2-86F5-4127E4A95283}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8B3DD29A-9904-45C4-89D5-BD767FBCB6F8}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Minsk.Core", "src\shared\Minsk.Core\Minsk.Core.csproj", "{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +34,18 @@ Global
 		{A5E27C6E-623B-46D2-86F5-4127E4A95283}.Release|x64.Build.0 = Release|Any CPU
 		{A5E27C6E-623B-46D2-86F5-4127E4A95283}.Release|x86.ActiveCfg = Release|Any CPU
 		{A5E27C6E-623B-46D2-86F5-4127E4A95283}.Release|x86.Build.0 = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|x64.Build.0 = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Debug|x86.Build.0 = Debug|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|x64.ActiveCfg = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|x64.Build.0 = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|x86.ActiveCfg = Release|Any CPU
+		{03D1DC8A-A865-4BE6-BE47-D3F1D794DDBD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
