--- mc/CodeAnalysis/Binding/BoundBinaryOperatorKind.cs
@@ -1,14 +1,0 @@
-﻿namespace mc.CodeAnalysis.Binding
-{
-    internal enum BoundBinaryOperatorKind
-    {
-        Addition,
-        Subtraction,
-        Multiplication,
-        Division,
-        LogicalAndAlso,
-        LogicalOrElse,
-        Equals,
-        NotEquals
-    }
-}
