--- mc/CodeAnalysis/Binding/BoundNode.cs
@@ -1,7 +1,0 @@
-﻿namespace mc.CodeAnalysis.Binding
-{
-    internal abstract class BoundNode
-    {
-        public abstract BoundNodeKind Kind { get;}
-    }
-}
