--- mc/CodeAnalysis/Binding/BoundNodeKind.cs
@@ -1,9 +1,0 @@
-﻿namespace mc.CodeAnalysis.Binding
-{
-    internal enum BoundNodeKind
-    {
-        UnaryExpression,
-        BinaryExpression,
-        LiteralExpression
-    }
-}
