--- mc/CodeAnalysis/Binding/BoundUnaryOperator.cs
@@ -1,39 +1,0 @@
-﻿using System;
-using System.Linq;
-using mc.CodeAnalysis.Syntax;
-
-namespace mc.CodeAnalysis.Binding
-{
-    internal sealed class BoundUnaryOperator
-    {
-        public SyntaxKind SyntaxKind { get; }
-        public BoundUnaryOperatorKind Kind { get; }
-        public Type OperandType { get; }
-        public Type Type { get; }
-
-        public BoundUnaryOperator(SyntaxKind syntaxKind, BoundUnaryOperatorKind kind, Type operandType) :
-            this(syntaxKind, kind, operandType, operandType)
-        {
-        }
-
-        public BoundUnaryOperator(SyntaxKind syntaxKind, BoundUnaryOperatorKind kind, Type operandType, Type resultType)
-        {
-            SyntaxKind = syntaxKind;
-            Kind = kind;
-            OperandType = operandType;
-            Type = resultType;
-        }
-
-        private static readonly BoundUnaryOperator[] _operators =
-        {
-            new BoundUnaryOperator(SyntaxKind.BangToken, BoundUnaryOperatorKind.LogicalNegation, typeof(bool)),
-            new BoundUnaryOperator(SyntaxKind.PlusToken, BoundUnaryOperatorKind.Identity, typeof(int)),
-            new BoundUnaryOperator(SyntaxKind.MinusToken, BoundUnaryOperatorKind.Negation, typeof(int)),
-        };
-
-        public static BoundUnaryOperator Bind(SyntaxKind syntaxKind, Type operandType)
-        {
-            return _operators.FirstOrDefault(op => op.SyntaxKind == syntaxKind && op.OperandType == operandType);
-        }
-    }
-}
