--- mc/CodeAnalysis/Binding/BoundUnaryOperatorKind.cs
@@ -1,9 +1,0 @@
-﻿namespace mc.CodeAnalysis.Binding
-{
-    internal enum BoundUnaryOperatorKind
-    {
-        Identity,
-        Negation,
-        LogicalNegation
-    }
-}
