--- mc/CodeAnalysis/Syntax/ExpressionSyntax.cs
@@ -1,7 +1,0 @@
-﻿namespace mc.CodeAnalysis.Syntax
-{
-    public abstract class ExpressionSyntax : SyntaxNode
-    {
-
-    }
-}
