--- mc/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,33 +1,0 @@
-﻿namespace mc.CodeAnalysis.Syntax
-{
-    public enum SyntaxKind
-    {
-        // Tokens
-        BadToken,
-        EofToken,
-        WhitespaceToken,
-        NumberToken,
-        PlusToken,
-        MinusToken,
-        StarToken,
-        SlashToken,
-        OpenParensToken,
-        CloseParensToken,
-        BangToken,
-        AmpersandAmpersandToken,
-        PipePipeToken,
-        EqualsEqualsToken,
-        BangEqualsToken,
-        Identifier,
-
-        // Expressions
-        LiteralExpression,
-        BinaryExpression,
-        ParenthesizedExpression,
-        UnaryExpression,
-
-        // Keywords
-        FalseKeyword,
-        TrueKeyword
-    }
-}
