--- mc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace mc.CodeAnalysis.Syntax
-{
-    public abstract class SyntaxNode
-    {
-        public abstract SyntaxKind Kind { get; }
-
-        public abstract IEnumerable<SyntaxNode> GetChildren();
-    }
-}
