--- mc/CodeAnalysis/Syntax/SyntaxToken.cs
@@ -1,29 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-
-namespace mc.CodeAnalysis.Syntax
-{
-    public class SyntaxToken : SyntaxNode
-    {
-        public SyntaxToken(SyntaxKind kind, int position, string text, object value)
-        {
-            Kind = kind;
-            Position = position;
-            Text = text;
-            Value = value;
-        }
-
-
-        public int Position { get; }
-
-        public string Text { get; }
-
-        public object Value { get; }
-        public override SyntaxKind Kind { get; }
-
-        public override IEnumerable<SyntaxNode> GetChildren()
-        {
-            return Enumerable.Empty<SyntaxNode>();
-        }
-    }
-}
