+++ src/app/Minsk.ConsoleApp/Minsk.ConsoleApp.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\shared\Minsk.Core\Minsk.Core.csproj" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <AssemblyName>Minsk.ConsoleApp</AssemblyName>
+    <RootNamespace>Minsk.ConsoleApp</RootNamespace>
+  </PropertyGroup>
+
+</Project>
