+++ src/shared/Minsk.Core/CodeAnalysis/Binding/Binder.cs
@@ -1,0 +1,64 @@
+﻿using System;
+using System.Collections.Generic;
+using Minsk.Core.CodeAnalysis.Syntax;
+
+namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal sealed class Binder
+    {
+        public List<string> Diagnostics { get; }
+
+        public Binder()
+        {
+            Diagnostics = new List<string>();
+        }
+
+        public BoundExpression BindExpression(ExpressionSyntax syntax)
+        {
+            switch (syntax.Kind)
+            {
+                case SyntaxKind.LiteralExpression:
+                    return BindLiteralExpression((LiteralExpressionSyntax)syntax);
+                case SyntaxKind.UnaryExpression:
+                    return BindUnaryExpression((UnaryExpressionSyntax)syntax);
+                case SyntaxKind.BinaryExpression:
+                    return BindBinaryExpression((BinaryExpressionSyntax)syntax);
+                case SyntaxKind.ParenthesizedExpression:
+                    return BindExpression(((ParenthesizedExpressionSyntax)syntax).Expression);
+                default:
+                    throw new Exception($"Unexpected syntax {syntax.Kind}");
+            }
+        }
+
+        private BoundExpression BindLiteralExpression(LiteralExpressionSyntax syntax)
+        {
+            var value = syntax.Value ?? 0;
+            return new BoundLiteralExpression(value);
+        }
+
+        private BoundExpression BindUnaryExpression(UnaryExpressionSyntax syntax)
+        {
+            var boundOperand = BindExpression(syntax.Operand);
+            var boundOperator = BoundUnaryOperator.Bind(syntax.OperatorToken.Kind, boundOperand.Type);
+
+            if (boundOperator != null) return new BoundUnaryExpression(boundOperator, boundOperand);
+
+            Diagnostics.Add($"Unary operator '{syntax.OperatorToken.Text}' is not defined for type {boundOperand.Type}");
+            return boundOperand;
+
+        }
+
+        private BoundExpression BindBinaryExpression(BinaryExpressionSyntax syntax)
+        {
+            var boundLeft = BindExpression(syntax.Left);
+            var boundRight = BindExpression(syntax.Right);
+            var boundOperator = BoundBinaryOperator.Bind(syntax.OperatorToken.Kind, boundLeft.Type, boundRight.Type);
+
+            if (boundOperator != null) return new BoundBinaryExpression(boundLeft, boundOperator, boundRight);
+
+            Diagnostics.Add($"Binary operator '{syntax.OperatorToken.Text}' is not defined for type {boundLeft.Type} and {boundRight.Type}");
+            return boundLeft;
+
+        }
+    }
+}
