+++ src/shared/Minsk.Core/CodeAnalysis/Binding/BoundBinaryOperatorKind.cs
@@ -1,0 +1,14 @@
+﻿namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal enum BoundBinaryOperatorKind
+    {
+        Addition,
+        Subtraction,
+        Multiplication,
+        Division,
+        LogicalAndAlso,
+        LogicalOrElse,
+        Equals,
+        NotEquals
+    }
+}
