+++ src/shared/Minsk.Core/CodeAnalysis/Binding/BoundExpression.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal abstract class BoundExpression : BoundNode
+    {
+        public abstract Type Type { get; }
+    }
+}
