+++ src/shared/Minsk.Core/CodeAnalysis/Binding/BoundNode.cs
@@ -1,0 +1,7 @@
+﻿namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal abstract class BoundNode
+    {
+        public abstract BoundNodeKind Kind { get;}
+    }
+}
