+++ src/shared/Minsk.Core/CodeAnalysis/Binding/BoundNodeKind.cs
@@ -1,0 +1,9 @@
+﻿namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal enum BoundNodeKind
+    {
+        UnaryExpression,
+        BinaryExpression,
+        LiteralExpression
+    }
+}
