+++ src/shared/Minsk.Core/CodeAnalysis/Binding/BoundUnaryOperatorKind.cs
@@ -1,0 +1,9 @@
+﻿namespace Minsk.Core.CodeAnalysis.Binding
+{
+    internal enum BoundUnaryOperatorKind
+    {
+        Identity,
+        Negation,
+        LogicalNegation
+    }
+}
