+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/ExpressionSyntax.cs
@@ -1,0 +1,7 @@
+﻿namespace Minsk.Core.CodeAnalysis.Syntax
+{
+    public abstract class ExpressionSyntax : SyntaxNode
+    {
+
+    }
+}
