+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/LiteralExpressionSyntax.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections.Generic;
+
+namespace Minsk.Core.CodeAnalysis.Syntax
+{
+    public sealed class LiteralExpressionSyntax : ExpressionSyntax
+    {
+        public LiteralExpressionSyntax(SyntaxToken literalToken) : this(literalToken, literalToken.Value)
+        {
+        }
+
+        public LiteralExpressionSyntax(SyntaxToken literalToken, object value)
+        {
+            LiteralToken = literalToken;
+            Value = value;
+        }
+
+        public override SyntaxKind Kind => SyntaxKind.LiteralExpression;
+        public override IEnumerable<SyntaxNode> GetChildren()
+        {
+            yield return LiteralToken;
+        }
+
+        public SyntaxToken LiteralToken { get; }
+        public object Value { get; }
+    }
+}
