+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,0 +1,33 @@
+﻿namespace Minsk.Core.CodeAnalysis.Syntax
+{
+    public enum SyntaxKind
+    {
+        // Tokens
+        BadToken,
+        EofToken,
+        WhitespaceToken,
+        NumberToken,
+        PlusToken,
+        MinusToken,
+        StarToken,
+        SlashToken,
+        OpenParensToken,
+        CloseParensToken,
+        BangToken,
+        AmpersandAmpersandToken,
+        PipePipeToken,
+        EqualsEqualsToken,
+        BangEqualsToken,
+        Identifier,
+
+        // Expressions
+        LiteralExpression,
+        BinaryExpression,
+        ParenthesizedExpression,
+        UnaryExpression,
+
+        // Keywords
+        FalseKeyword,
+        TrueKeyword
+    }
+}
