+++ src/shared/Minsk.Core/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Minsk.Core.CodeAnalysis.Syntax
+{
+    public abstract class SyntaxNode
+    {
+        public abstract SyntaxKind Kind { get; }
+
+        public abstract IEnumerable<SyntaxNode> GetChildren();
+    }
+}
